refactor: extract dead_letter ops into diesel_common
