+++ crates/taskito-core/src/storage/diesel_common/mod.rs
@@ -3,12 +3,14 @@
 //! These macros generate the identical method implementations that both
 //! Diesel-backed storage backends share, avoiding code duplication.
 
+mod dead_letter;
 mod jobs;
 mod locks;
 mod logs;
 mod metrics;
 pub(crate) mod migrations;
 
+pub(crate) use dead_letter::impl_diesel_dead_letter_ops;
 pub(crate) use jobs::impl_diesel_job_ops;
 pub(crate) use locks::impl_diesel_lock_ops;
 pub(crate) use logs::impl_diesel_log_ops;
